fix(repo): update cookie type to be tighter to string OR string[]
refactor(repo): update urlboxoptions types

refactor(repo): split classes into their own files
