+++ Urlbox/Urlbox/Urlbox.cs
@@ -278,181 +278,5 @@ namespace Screenshots
 
     }
 
-    /// <summary>
-    /// Interface for Urlbox response types.
-    /// Allows one response type for makeUrlboxPostRequest which can then
-    /// be cast to the specific /sync or /async response
-    /// Implementations represent either synchronous or asynchronous responses.
-    /// </summary>
-    public interface IUrlboxResponse
-    {
-    }
 
-    /// <summary>
-    /// Represents a synchronous Urlbox response.
-    /// </summary>
-    public class SyncUrlboxResponse : IUrlboxResponse
-    {
-        public string RenderUrl { get; set; }
-        public int Size { get; set; }
-    }
-
-    /// <summary>
-    /// Represents an asynchronous Urlbox response.
-    /// </summary>
-    public class AsyncUrlboxResponse : IUrlboxResponse
-    {
-        public string Status { get; set; }
-        public string RenderId { get; set; }
-        public string StatusUrl { get; set; }
-    }
-
-
-    /// <summary>
-    /// Initializes a new instance of the UrlboxOptions. These are used as part of any Urlbox method which requires render options.
-    /// </summary>
-    /// <exception cref="ArgumentException">Thrown when the Url OR Html option isn't passed in on init.</exception>
-    public class UrlboxOptions
-    {
-
-        public UrlboxOptions(string url = null, string html = null)
-        {
-            if (string.IsNullOrEmpty(url) && string.IsNullOrEmpty(html))
-            {
-                throw new ArgumentException("Either of options 'url' or 'html' must be provided.");
-            }
-            Url = url;
-            Html = html;
-        }
-
-        public string Url { get; set; }
-        public string Html { get; set; }
-        public string Format { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
-        public bool FullPage { get; set; }
-        public string Selector { get; set; }
-        public string Clip { get; set; }
-        public bool Gpu { get; set; }
-        public string ResponseType { get; set; }
-        public bool BlockAds { get; set; }
-        public bool HideCookieBanners { get; set; }
-        public bool ClickAccept { get; set; }
-        public string[] BlockUrls { get; set; }
-        public bool BlockImages { get; set; }
-        public bool BlockFonts { get; set; }
-        public bool BlockMedias { get; set; }
-        public bool BlockStyles { get; set; }
-        public bool BlockScripts { get; set; }
-        public bool BlockFrames { get; set; }
-        public bool BlockFetch { get; set; }
-        public bool BlockXhr { get; set; }
-        public bool BlockSockets { get; set; }
-        public string HideSelector { get; set; }
-        public string Js { get; set; }
-        public string Css { get; set; }
-        public bool DarkMode { get; set; }
-        public bool ReducedMotion { get; set; }
-        public bool Retina { get; set; }
-        public int ThumbWidth { get; set; }
-        public int ThumbHeight { get; set; }
-        public string ImgFit { get; set; }
-        public string ImgPosition { get; set; }
-        public string ImgBg { get; set; }
-        public int ImgPad { get; set; }
-        public int Quality { get; set; }
-        public bool Transparent { get; set; }
-        public int MaxHeight { get; set; }
-        public string Download { get; set; }
-        public string PdfPageSize { get; set; }
-        public string PdfPageRange { get; set; }
-        public int PdfPageWidth { get; set; }
-        public int PdfPageHeight { get; set; }
-        public string PdfMargin { get; set; }
-        public int PdfMarginTop { get; set; }
-        public int PdfMarginRight { get; set; }
-        public int PdfMarginBottom { get; set; }
-        public int PdfMarginLeft { get; set; }
-        public bool PdfAutoCrop { get; set; }
-        public double PdfScale { get; set; }
-        public string PdfOrientation { get; set; }
-        public bool PdfBackground { get; set; }
-        public bool DisableLigatures { get; set; }
-        public string Media { get; set; }
-        public bool PdfShowHeader { get; set; }
-        public string PdfHeader { get; set; }
-        public bool PdfShowFooter { get; set; }
-        public string PdfFooter { get; set; }
-        public bool Readable { get; set; }
-        public bool Force { get; set; }
-        public string Unique { get; set; }
-        public int Ttl { get; set; }
-        public string Proxy { get; set; }
-        public string Header { get; set; }
-        public object Cookie { get; set; }
-        public string UserAgent { get; set; }
-        public string Platform { get; set; }
-        public string AcceptLang { get; set; }
-        public string Authorization { get; set; }
-        public string Tz { get; set; }
-        public string EngineVersion { get; set; }
-        public int Delay { get; set; }
-        public int Timeout { get; set; }
-        public string WaitUntil { get; set; }
-        public string WaitFor { get; set; }
-        public string WaitToLeave { get; set; }
-        public int WaitTimeout { get; set; }
-        public bool FailIfSelectorMissing { get; set; }
-        public bool FailIfSelectorPresent { get; set; }
-        public bool FailOn4xx { get; set; }
-        public bool FailOn5xx { get; set; }
-        public string ScrollTo { get; set; }
-        public string Click { get; set; }
-        public string ClickAll { get; set; }
-        public string Hover { get; set; }
-        public string BgColor { get; set; }
-        public bool DisableJs { get; set; }
-        public string FullPageMode { get; set; }
-        public bool FullWidth { get; set; }
-        public bool AllowInfinite { get; set; }
-        public bool SkipScroll { get; set; }
-        public bool DetectFullHeight { get; set; }
-        public int MaxSectionHeight { get; set; }
-        public string ScrollIncrement { get; set; }
-        public int ScrollDelay { get; set; }
-        public string Highlight { get; set; }
-        public string HighlightFg { get; set; }
-        public string HighlightBg { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
-        public int Accuracy { get; set; }
-        public bool UseS3 { get; set; }
-        public string S3Path { get; set; }
-        public string S3Bucket { get; set; }
-        public string S3Endpoint { get; set; }
-        public string S3Region { get; set; }
-        public string CdnHost { get; set; }
-        public string S3StorageClass { get; set; }
-    }
-
-
-    /// <summary>
-    /// A custom naming policy for converting property names from PascalCase to snake_case
-    /// when serializing JSON.
-    /// 
-    /// <remarks>
-    /// This JsonNamingPolicy is included by default in .NET 8.0 (JsonNamingPolicy.SnakeCaseLower).
-    /// However, a custom implementation has been made here to maintain compatibility with .NET 6.0,
-    /// which is still under Long-Term Support (LTS). Keeping the SDK at 6.0 ensures broader accessibility 
-    /// for audiences still using this version.
-    /// </remarks>
-    /// </summary>
-    public class SnakeCaseNamingPolicy : JsonNamingPolicy
-    {
-        public override string ConvertName(string name)
-        {
-            // Convert PascalCase to snake_case
-            return string.Concat(name.Select((x, i) => i > 0 && char.IsUpper(x) ? "_" + x.ToString() : x.ToString())).ToLower();
-        }
-    }
 }
